+++ pkg/interfaces/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### New Features
+
+- Added `locked__end` to `IVotingEscrow`.
+- Added `IL2LayerZeroDelegation`.
+
 ## 0.4.0 (2023-03-15)
 
 - Added `IERC4626`.